core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sunday"
version = "0.1.0"
description = "SUNDAY — Simply Unique Natural Daily Assistant for YOU"
readme = "../README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "SUNDAY Team" }]

dependencies = [
    # API Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",

    # LLM
    "litellm>=1.50.0",
    "httpx>=0.27.0",

    # Voice
    "faster-whisper>=1.1.0",
    "piper-tts>=1.2.0",
    "silero-vad>=5.1.2",
    "sounddevice>=0.5.0",
    "numpy>=1.26.0",

    # Database
    "aiosqlite>=0.20.0",
    "chromadb>=0.5.0",

    # Config & Utils
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
sunday = "sunday.main:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["sunday"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
